refactor MetabotChatSuggestions
