+++ enterprise/frontend/src/embedding-sdk-bundle/components/public/MetabotQuestion/MetabotSidebar.tsx
@@ -2,6 +2,7 @@ import { Stack } from "metabase/ui";
 
 import { MetabotChatHistory } from "./MetabotChatHistory";
 import { MetabotChatInput } from "./MetabotChatInput";
+import { MetabotChatSuggestions } from "./MetabotChatSuggestions";
 import S from "./MetabotQuestion.module.css";
 import { SidebarHeader } from "./SidebarHeader";
 
@@ -10,7 +11,11 @@ export function MetabotSidebar() {
     <Stack className={S.sidebar}>
       <SidebarHeader />
       <MetabotChatHistory />
-      <MetabotChatInput />
+
+      <Stack gap={0}>
+        <MetabotChatSuggestions />
+        <MetabotChatInput />
+      </Stack>
     </Stack>
   );
 }
